+++ src/terminal/Functions.h
@@ -295,6 +295,7 @@ constexpr inline auto TBC         = detail::CSI(std::nullopt, 0, 1, std::nullopt
 constexpr inline auto VPA         = detail::CSI(std::nullopt, 0, 1, std::nullopt, 'd', VTType::VT100, "VPA", "Vertical Position Absolute");
 constexpr inline auto WINMANIP    = detail::CSI(std::nullopt, 1, 3, std::nullopt, 't', VTType::VT525, "WINMANIP", "Window Manipulation");
 constexpr inline auto XTSMGRAPHICS= detail::CSI('?', 2, 4, std::nullopt, 'S', VTType::VT525 /*XT*/, "XTSMGRAPHICS", "Setting/getting Sixel/ReGIS graphics settings.");
+constexpr inline auto CAPTURE     = detail::CSI('>', 0, 2, std::nullopt, 't', VTType::VT525 /*Extension*/, "CAPTURE", "Report screen buffer capture.");
 
 // DCS functions
 constexpr inline auto STP         = detail::DCS(std::nullopt, 0, 0, '$', 'p', VTType::VT525, "STP", "Set Terminal Profile");
@@ -370,6 +371,7 @@ inline auto const& functions() noexcept
 
             // CSI
             ANSISYSSC,
+            CAPTURE,
             CBT,
             CHA,
             CHT,